+++ packages/web3-core/test/unit/web3_config.test.ts
@@ -33,6 +33,7 @@ const defaultConfig = {
 		useRpcCallSpecification: false,
 	},
 	handleRevert: false,
+	contractDataInputFill: 'input',
 	maxListenersWarningThreshold: 100,
 	transactionBlockTimeout: 50,
 	transactionConfirmationBlocks: 24,
@@ -41,7 +42,7 @@ const defaultConfig = {
 	transactionReceiptPollingInterval: undefined,
 	transactionSendTimeout: 750 * 1000,
 	transactionConfirmationPollingInterval: undefined,
-	defaultTransactionType: '0x0',
+	defaultTransactionType: '0x2',
 	defaultMaxPriorityFeePerGas: toHex(2500000000),
 };
 const setValue = {