+++ packages/web3-eth-contract/test/integration/local_account/contract_overloaded_methods.test.ts
@@ -20,7 +20,8 @@ import Web3 from 'web3';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Web3Account } from 'web3-eth-accounts';
 import { utf8ToHex } from 'web3-utils';
-import { Contract, EventLog } from '../../../src';
+import { EventLog } from 'web3-types';
+import { Contract } from '../../../src';
 import { ERC721TokenAbi, ERC721TokenBytecode } from '../../shared_fixtures/build/ERC721Token';
 import { getSystemTestProvider, createLocalAccount } from '../../fixtures/system_test_utils';
 import { toUpperCaseHex } from '../../shared_fixtures/utils';