+++ packages/web3-eth-accounts/CHANGELOG.md
@@ -130,4 +130,11 @@ Documentation:
 
 -   Dependencies updated
 
+## [4.0.6]
+
+### Fixed
+
+-   Fixed "The `r` and `s` returned by `sign` to does not always consist of 64 characters" (#6411)
+
+
 ## [Unreleased]