+++ packages/web3-net/CHANGELOG.md
@@ -129,4 +129,10 @@ Documentation:
 
 -   Dependencies updated
 
+## [4.0.6]
+
+### Changed
+
+-   Dependencies updated
+
 ## [Unreleased]