+++ packages/web3-types/CHANGELOG.md
@@ -157,4 +157,15 @@ Documentation:
 
 -   Dependencies updated
 
+## [1.2.0]
+
+### Added
+
+-   add `asEIP1193Provider` to `Web3BaseProvider` so every inherited class can have the returned value of `request` method, fully compatible with EIP-1193. (#6407)
+
+
 ## [Unreleased]
+
+### Added
+
+-   Interface `EventLog` was added. (#6410)