+++ packages/web3-utils/CHANGELOG.md
@@ -152,4 +152,10 @@ Documentation:
 
 -   Dependencies updated
 
+## [4.0.6]
+
+### Fixed
+
+-   `soliditySha3()` with BigInt support
+
 ## [Unreleased]